+++ torch/csrc/api/include/torch/data/dataloader_options.h
@@ -8,19 +8,9 @@
 
 namespace torch {
 namespace data {
-namespace detail {
-struct FullDataLoaderOptions;
-} // namespace detail
-} // namespace data
-} // namespace torch
-
-namespace torch {
-namespace data {
 
 /// Options to configure a `DataLoader`.
 struct DataLoaderOptions {
-  friend struct torch::data::detail::FullDataLoaderOptions;
-
   DataLoaderOptions() = default;
   /* implicit */ DataLoaderOptions(size_t batch_size)
       : batch_size_(batch_size) {}
