+++ torch/csrc/api/include/torch/data/datasets/stateful.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <torch/data/datasets/base.h>
+#include <torch/data/datasets/map.h>
 #include <torch/data/example.h>
 
 #include <torch/types.h>
